backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==1.4.47
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis>=2.10.6,<5.0.0
aioredis==2.0.1

# Data Validation
pydantic>=1.10.12,<2.0.0
pydantic[email]>=1.10.12,<2.0.0
email-validator==2.1.0

# RSS/Feed Parsing
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP Client
httpx>=0.18,<0.25
aiohttp==3.9.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Task Scheduling
apscheduler==3.10.4
celery==5.3.4
flower==2.0.1

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# OAuth2 Providers
authlib==1.2.1
httpx-oauth==0.13.0

# File Processing
python-magic==0.4.27
openpyxl==3.1.2
pandas==2.1.3

# Security
cryptography==41.0.7

# Monitoring and Logging
sentry-sdk==1.38.0
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
ipython==8.17.2
ipdb==0.13.13

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings[python]==0.24.0

# Performance
orjson==3.9.10
ujson==5.8.0
msgpack==1.0.7

# Caching
aiocache==0.12.2
cachetools==5.3.2

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Data Processing
numpy==1.26.2
scipy==1.11.4

# Image Processing (for future avatars/thumbnails)
pillow==10.1.0

# XML Processing
defusedxml==0.7.1

# Markdown Processing
markdown==3.5.1
markdown2==2.4.12

# Search
elasticsearch==8.11.0  # Optional for advanced search
whoosh==2.7.4  # Alternative lightweight search

# Compression
brotli==1.1.0

# Profiling (Development)
py-spy==0.3.14
memory-profiler==0.61.0

# API Documentation
pydantic-openapi-schema==1.5.1
fastapi-pagination==0.12.13

# Background Tasks Alternative
dramatiq[redis,watch]==1.15.0
arq==0.25.0

# Health Checks
fastapi-health==0.4.0

# Feature Flags
flipper-client==1.3.2

# GraphQL Support (Optional)
strawberry-graphql[fastapi]==0.215.1

# Migrations
yoyo-migrations==8.2.0

# CLI Tools
typer==0.7.0
rich==13.7.0
click==8.1.7

# Timezone
tzdata==2023.3

# URL Validation
validators==0.22.0

# HTML Sanitization
bleach==6.1.0

# Natural Language Processing (Optional for advanced search)
spacy==3.7.2
textblob==0.17.1

# Async Support
anyio>=3.7.1,<4.0.0
asyncer==0.0.2

# Configuration Management
dynaconf==3.2.4

# Database Connection Pooling
sqlalchemy-utils==0.41.1
databases==0.8.0

# API Client Generation
openapi-python-client==0.14.0  # downgraded for compatibility with pydantic 1.x

# Code Quality
pylint==3.0.3
bandit==1.7.5
safety==3.0.1